pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "rule-generator"
version = "0.1.0"
description = "AI-assisted Loader ETL rule generator"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "click>=8.1",
    "questionary>=2.0",
    "openai>=1.0",
    "anthropic>=0.25",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "sqlparse>=0.5",
]

[project.scripts]
rule-gen = "src.cli.main:cli"
